requirements.txt
asgiref==3.5.0
astroid==2.11.6
Babel==2.9.1
backports.zoneinfo==0.2.1
black==22.6.0
branca==0.4.2
certifi==2021.10.8
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
decorator==5.1.1
dill==0.3.5.1
distlib==0.3.6
dj-database-url==0.5.0
Django==3.2.13
django-appconf==1.0.5
django-countries==7.2.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.9.0
django-extensions==3.1.5
django-extra-views==0.14.0
django-octicons==1.0.2
django-phonenumber-field==6.1.0
django-select2==7.10.0
django-widget-tweaks==1.4.12
dodgy==0.2.1
Faker==13.0.0
filelock==3.9.0
flake8==4.0.1
flake8-polyfill==1.0.2
folium==0.12.1.post1
future==0.18.2
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
identify==2.5.16
idna==3.3
isort==5.10.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.22.2
pathspec==0.9.0
pep8-naming==0.10.0
phonenumbers==8.12.43
platformdirs==2.5.2
pre-commit==3.0.1
prospector==1.7.7
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
pylint==2.14.3
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
ratelim==0.1.6
requests==2.27.1
requirements-detector==0.7
setoptconf-tmp==0.3.1
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typing_extensions==4.2.0
urllib3==1.26.8
virtualenv==20.17.1
wrapt==1.14.1
